+++ src/BiliLite.UWP/Modules/Player/InteractionVideoVM.cs
@@ -55,7 +55,7 @@ namespace BiliLite.Modules.Player
             set { _select = value; DoPropertyChanged("Select"); }
         }
 
-        public async Task GetNodes(int edge_id = 0)
+        public async Task GetNodes(long edge_id = 0)
         {
             try
             {
@@ -126,7 +126,7 @@ namespace BiliLite.Modules.Player
   
     public class InteractionEdgeInfoChoiceModel
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string platform_action { get; set; }
         public string native_action { get; set; }
         public string condition { get; set; }
@@ -139,7 +139,7 @@ namespace BiliLite.Modules.Player
 
     public class InteractionEdgeInfoQuestionModel
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public int type { get; set; }
         public int start_time_r { get; set; }
         public int duration { get; set; }
